refactor: usage of auth and identity resolvers in orchestrator interceptor test add: runtime plugin for anonymous auth fix: clippy lints for explicit lifetimes remove: test-utils endpoints and auth modules
