+++ rust-runtime/aws-smithy-runtime-api/Cargo.toml
@@ -12,6 +12,7 @@ repository = "https://github.com/awslabs/smithy-rs"
 [features]
 default = []
 http-auth = ["dep:zeroize"]
+anonymous-auth = []
 
 [dependencies]
 aws-smithy-async = { path = "../aws-smithy-async" }
