+++ rust-runtime/aws-smithy-runtime-api/src/client/runtime_plugin.rs
@@ -6,6 +6,9 @@
 use crate::client::interceptors::Interceptors;
 use crate::config_bag::ConfigBag;
 
+#[cfg(feature = "anonymous-auth")]
+pub mod anonymous_auth;
+
 pub type BoxError = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 pub trait RuntimePlugin {
