+++ rust-runtime/aws-smithy-runtime/Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/awslabs/smithy-rs"
 
 [features]
 http-auth = ["aws-smithy-runtime-api/http-auth"]
+anonymous-auth = ["aws-smithy-runtime-api/anonymous-auth"]
 test-util = ["dep:aws-smithy-protocol-test"]
 
 [dependencies]
