+++ rust-runtime/aws-smithy-runtime/src/client/test_util.rs
@@ -3,8 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-pub mod auth;
 pub mod connector;
 pub mod deserializer;
-pub mod endpoints;
 pub mod serializer;
